+++ src/kernel/Multiplier.java
@@ -27,7 +27,7 @@ public class Multiplier {
 	void activate() {
 		time--;
 		if (time == 0) {
-			double ans;
+			float ans;
 			if (instr.op == OP.MUL) {
 				ans = main.fp.get(instr.src1.ordinal()) * main.fp.get(instr.src2.ordinal());
 			}
